refactor(apps/esb/permission): remove esb/permission in edition ee
